database: move database setup into helper package
...to be used from internal/server, too.
